Moved files to the root folder.
