+++ newCrawler/parseHelper.js
